Ready add to BD

+++ repositories/src/main/scala/repositories/BaseRepository.scala
@@ -2,6 +2,8 @@ package repositories
 
 import model.HasId
 import slick.jdbc.PostgresProfile.api._
+import tables.BookTable
+
 import collection.immutable
 import scala.concurrent.Future
 
@@ -9,7 +11,7 @@ import scala.concurrent.Future
 trait BaseTable[A <: HasId] {
   self: Table[A] =>
 
-  val id: Rep[Int]
+  val id: Rep[Long]
 }
 
 abstract class CrudRepository[A](query: TableQuery[_ <: Table[A]]) {
@@ -20,12 +22,12 @@ abstract class CrudRepository[A](query: TableQuery[_ <: Table[A]]) {
   def insert(items: immutable.Seq[A]): Future[Option[Int]] = db.run(query ++= items)
 }
 
-abstract class BaseRepository[A <: HasId](query: TableQuery[_ <: Table[A] with BaseTable[A]]) extends CrudRepository[A](query) {
+abstract class BaseRepository[A <: HasId](query: TableQuery[_ <: Table[A] with BaseTable[A]]) extends CrudRepository[A](query){
   def db: Database
 
-  def findById(id: Int): Future[Option[A]] = db.run(query.filter(_.id === id).result.headOption)
+  def findById(id: Long): Future[Option[A]] = db.run(query.filter(_.id === id).result.headOption)
 
   def update(item: A): Future[Int] = db.run(query.filter(_.id === item.id).update(item))
 
-  def deleteById(id: Int): Future[Int] = db.run(query.filter(_.id === id).delete)
+  def deleteById(id: Long): Future[Int] = db.run(query.filter(_.id === id).delete)
 }
