+++ repositories/src/main/scala/tables/BookGenreTable.scala
@@ -4,8 +4,8 @@ import model._
 import slick.jdbc.PostgresProfile.api._
 
 class BookGenreTable(tag: Tag) extends Table[BookGenre](tag, "book_genre"){
-  val genreId = column[Int]("gen_id")
-  val bookId = column[Int]("bk_id")
+  val genreId = column[Long]("gen_id")
+  val bookId = column[Long]("bk_id")
 
   def * = (genreId, bookId) <> ((BookGenre.apply _).tupled, BookGenre.unapply)
